Support Python 2.6
According to the Python documentation:

	Changed in version 2.7: The positional argument specifiers can be
	omitted, so '{} {}' is equivalent to '{0} {1}'.

In general, it is a sign of bad development practices to require upgrades
on the user part, especially when it is very easy not to require that.
(See also https://en.wikipedia.org/wiki/Postel%27s_law).

So just provide the positional argument specifiers and be done with it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>